Switch to u64

+++ mqtt/mqtt-bridge/src/persist/waking_state/ring_buffer/block.rs
@@ -67,7 +67,7 @@ impl BlockHeaderV1 {
         }
     }
 
-    pub fn hint(&self) -> u32 {
+    pub fn hint(&self) -> u64 {
         self.hint
     }
